subprojects available

+++ cpppm/cli.py
@@ -20,16 +20,18 @@ from .project import Project
               type=click.Choice(['Debug', 'Release'], case_sensitive=True), help="Build type, Debug or Release.")
 @click.pass_context
 def cli(ctx, verbose, out_directory, clean, setting, build_type):
+    if not Project.current_project.is_root:
+        return
     if clean:
-        out_directory = Path(out_directory) if out_directory else Project.root_project.build_path
+        out_directory = Path(out_directory) if out_directory else Project._root_project.build_path
         if out_directory.exists():
             shutil.rmtree(out_directory)
     Project.settings = list(setting)
     Project.build_type = build_type
     if out_directory:
         Project.set_build_path(Path(out_directory))
-    Project.root_project.build_path.mkdir(exist_ok=True)
-    if not Project.root_project.build_path.exists():
+    Project.current_project.build_path.mkdir(exist_ok=True)
+    if not Project.current_project.build_path.exists():
         raise RuntimeError('Failed to create build directory: {build_directory}')
     if verbose:
         logging.basicConfig(level=logging.DEBUG)
@@ -45,7 +47,7 @@ def __cpppm_event__():
 @cli.command('install-requirements')
 def install_requirements():
     """Generates conan/CMake stuff."""
-    Project.root_project.install_requirements()
+    Project._root_project.install_requirements()
 
 
 @cli.command()
@@ -53,7 +55,7 @@ def install_requirements():
 def generate(ctx):
     """Generates conan/CMake stuff."""
     ctx.invoke(install_requirements)
-    Project.root_project.generate()
+    Project._root_project.generate()
 
 
 @cli.command()
@@ -70,10 +72,10 @@ def build(ctx, target):
     """Builds the project."""
     source_dir = Path(sys.argv[0]).parent
     click.echo(f"Source directory: {str(source_dir.absolute())}")
-    click.echo(f"Build directory: {str(Project.root_project.build_path.absolute())}")
-    click.echo(f"Project: {Project.root_project.name}")
+    click.echo(f"Build directory: {str(Project._root_project.build_path.absolute())}")
+    click.echo(f"Project: {Project._root_project.name}")
     ctx.invoke(configure)
-    rc = Project.root_project.build(target)
+    rc = Project._root_project.build(target)
     if rc != 0:
         click.echo(f'Build failed with return code: {rc}')
         exit(rc)
@@ -94,4 +96,4 @@ def test(ctx):
 def run(ctx, target, args):
     """Runs the given TARGET with given ARGS."""
     ctx.invoke(build)
-    Project.root_project.run(target, *args)
+    Project._root_project.run(target, *args)
